agrego lo avanzado en orden de compra. Marca, carga de cantidad real recibida

+++ src/dto/OrdenCompraRepuestoDTO.java
@@ -72,7 +72,7 @@ public class OrdenCompraRepuestoDTO {
 		this.habilitado = habilitado;
 	}
 
-	public float getCantidad_real() {
+	public int getCantidad_real() {
 		return cantidad_real;
 	}
